requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.3
cloudevents==1.2.0
deprecation==2.1.0
git+https://github.com/Rapptz/discord.py
git+https://github.com/phenom4n4n/TagScript
expr.py==0.3.0
Flask==2.0.2
functions-framework==3.0.0
google-api-core==2.3.0
google-auth==2.3.3
google-cloud-appengine-logging==1.1.0
google-cloud-audit-log==0.2.0
google-cloud-core==2.2.1
google-cloud-error-reporting==1.4.1
google-cloud-logging==2.7.0
googleapis-common-protos==1.54.0
grpc-google-iam-v1==0.12.3
grpcio==1.42.0
grpcio-status==1.42.0
gunicorn==20.1.0
idna==3.3
orjson==3.6.7
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
multidict==5.2.0
packaging==21.3
proto-plus==1.19.8
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyNaCl==1.4.0
pyparsing==3.0.6
requests==2.26.0
rsa==4.8
six==1.16.0
typing-extensions==4.0.1
urllib3==1.26.7
watchdog==1.0.2
Werkzeug==2.0.2
yarl==1.7.2
pymongo==4.0.1
dnspython==2.1.0
sentry-sdk==1.5.1
redis==4.1.0
fuzzywuzzy==0.18.0  